pyproject.toml
[tool.poetry]
name = "competencias_backend"
version = "0.1.0"
description = "Prototipo do backend do TCC"
authors = ["Chico Moura"]

[tool.poetry.dependencies]
python = "^3.10.6"
Django = "^4.1"
psycopg2-binary = "^2.9.3"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.13.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pytest-django = "^4.5.2"
mockito = "^1.3.5"
factory-boy = "^3.2.1"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
